+++ cakeshop-api/src/main/java/com/jpmorgan/cakeshop/model/NodeConfig.java
@@ -10,12 +10,12 @@ public class NodeConfig {
 
     private String identity;
     private Boolean committingTransactions;
-    private Integer networkId;
+    private Long networkId;
     private Integer logLevel;
     private String genesisBlock;
     private String extraParams;
 
-    public NodeConfig (String identity, Boolean mining, Integer networkid, Integer verbosity, String genesisBlock, String extraParams) {
+    public NodeConfig (String identity, Boolean mining, Long networkid, Integer verbosity, String genesisBlock, String extraParams) {
         this.identity = identity;
         this.committingTransactions = mining;
         this.networkId = networkid;
@@ -55,14 +55,14 @@ public class NodeConfig {
     /**
      * @return the networkId
      */
-    public Integer getNetworkId() {
+    public Long getNetworkId() {
         return networkId;
     }
 
     /**
      * @param networkId the networkId to set
      */
-    public void setNetworkid(Integer networkId) {
+    public void setNetworkid(Long networkId) {
         this.networkId = networkId;
     }
